完成邮箱各界面的功能实现，并尝试搓一个配置文件gui以更方便的在游戏内更改配置文件。

+++ src/main/java/cat/jiu/email/net/msg/MsgDeleteEmail.java
@@ -84,7 +84,7 @@ public class MsgDeleteEmail {
 					
 					for(long i : inbox.getEmailIDs()) {
 						Email email = inbox.getEmail(i);
-						if(email.isReceived() && !MinecraftForge.EVENT_BUS.post(new EmailDeleteEvent.Pre(inbox, i, false, true))) {
+						if(email.isReceived() && email.isRead() && !MinecraftForge.EVENT_BUS.post(new EmailDeleteEvent.Pre(inbox, i, false, true))) {
 							inbox.deleteEmail(i);
 						}
 						MinecraftForge.EVENT_BUS.post(new EmailDeleteEvent.Post(inbox, i, false, true));
